FLIZ-452 feat: /notification/session 페이지 네트워크 최적화
